requirements.txt
certifi==2022.9.14
charset-normalizer==2.0.12
click==8.1.3
contextlib2==21.6.0
docopt==0.6.2
idna==3.3
packaging==21.3
pip==22.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
pythonanywhere==0.10.1
PyYAML==6.0
requests==2.27.1
schema==0.7.5
setuptools==63.4.1
six==1.16.0
tabulate==0.8.9
typer==0.4.0
urllib3==1.26.8
wheel==0.37.1
